[#388] refactor: postService 수정
