--- src/main/java/com/jeju/nanaland/domain/common/service/PostCardService.java
@@ -1,9 +1,0 @@
-package com.jeju.nanaland.domain.common.service;
-
-import com.jeju.nanaland.domain.common.data.Language;
-import com.jeju.nanaland.domain.common.dto.PostCardDto;
-
-public interface PostCardService {
-
-  PostCardDto getPostCardDto(Long postId, Language language);
-}
