--- src/main/java/com/jeju/nanaland/domain/experience/service/ExperienceSearchService.java
@@ -1,22 +1,0 @@
-package com.jeju.nanaland.domain.experience.service;
-
-import com.jeju.nanaland.domain.common.data.Category;
-import com.jeju.nanaland.domain.common.entity.Post;
-import com.jeju.nanaland.domain.common.service.PostSearchService;
-import com.jeju.nanaland.domain.experience.repository.ExperienceRepository;
-import com.jeju.nanaland.global.exception.NotFoundException;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class ExperienceSearchService implements PostSearchService {
-
-  private final ExperienceRepository experienceRepository;
-
-  @Override
-  public Post getPost(Long postId, Category category) {
-    return experienceRepository.findById(postId)
-        .orElseThrow(() -> new NotFoundException("해당 게시물을 찾을 수 없습니다."));
-  }
-}
