+++ src/main/java/com/jeju/nanaland/domain/experience/service/ExperienceService.java
@@ -2,9 +2,14 @@ package com.jeju.nanaland.domain.experience.service;
 
 import static com.jeju.nanaland.domain.common.data.Category.EXPERIENCE;
 
+import com.jeju.nanaland.domain.common.data.Category;
 import com.jeju.nanaland.domain.common.data.Language;
+import com.jeju.nanaland.domain.common.data.PostCategory;
 import com.jeju.nanaland.domain.common.dto.ImageFileDto;
+import com.jeju.nanaland.domain.common.dto.PostCardDto;
+import com.jeju.nanaland.domain.common.entity.Post;
 import com.jeju.nanaland.domain.common.repository.ImageFileRepository;
+import com.jeju.nanaland.domain.common.service.PostService;
 import com.jeju.nanaland.domain.experience.dto.ExperienceCompositeDto;
 import com.jeju.nanaland.domain.experience.dto.ExperienceResponse.ExperienceDetailDto;
 import com.jeju.nanaland.domain.experience.dto.ExperienceResponse.ExperienceThumbnail;
@@ -21,6 +26,7 @@ import com.jeju.nanaland.global.exception.ErrorCode;
 import com.jeju.nanaland.global.exception.NotFoundException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -32,13 +38,29 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class ExperienceService {
+public class ExperienceService implements PostService {
 
   private final ExperienceRepository experienceRepository;
   private final FavoriteService favoriteService;
   private final ImageFileRepository imageFileRepository;
   private final SearchService searchService;
   private final ReviewRepository reviewRepository;
+
+  @Override
+  public Post getPost(Long postId, Category category) {
+    return experienceRepository.findById(postId)
+        .orElseThrow(() -> new NotFoundException("해당 게시물을 찾을 수 없습니다."));
+  }
+
+  @Override
+  public PostCardDto getPostCardDto(Long postId, Category category, Language language) {
+    PostCardDto postCardDto = experienceRepository.findPostCardDto(postId, language);
+    Optional.ofNullable(postCardDto)
+        .orElseThrow(() -> new NotFoundException("해당 게시물을 찾을 수 없습니다."));
+
+    postCardDto.setCategory(PostCategory.EXPERIENCE.toString());
+    return postCardDto;
+  }
 
   // 이색체험 리스트 조회
   public ExperienceThumbnailDto getExperienceList(MemberInfoDto memberInfoDto,
