+++ src/main/java/com/jeju/nanaland/domain/nature/service/NatureService.java
@@ -2,7 +2,13 @@ package com.jeju.nanaland.domain.nature.service;
 
 import static com.jeju.nanaland.domain.common.data.Category.NATURE;
 
+import com.jeju.nanaland.domain.common.data.Category;
+import com.jeju.nanaland.domain.common.data.Language;
+import com.jeju.nanaland.domain.common.data.PostCategory;
+import com.jeju.nanaland.domain.common.dto.PostCardDto;
+import com.jeju.nanaland.domain.common.entity.Post;
 import com.jeju.nanaland.domain.common.service.ImageFileService;
+import com.jeju.nanaland.domain.common.service.PostService;
 import com.jeju.nanaland.domain.favorite.service.FavoriteService;
 import com.jeju.nanaland.domain.member.dto.MemberResponse.MemberInfoDto;
 import com.jeju.nanaland.domain.nature.dto.NatureCompositeDto;
@@ -14,6 +20,7 @@ import com.jeju.nanaland.domain.search.service.SearchService;
 import com.jeju.nanaland.global.exception.ErrorCode;
 import com.jeju.nanaland.global.exception.NotFoundException;
 import java.util.List;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
@@ -24,12 +31,28 @@ import org.springframework.stereotype.Service;
 @Service
 @RequiredArgsConstructor
 @Slf4j
-public class NatureService {
+public class NatureService implements PostService {
 
   private final NatureRepository natureRepository;
   private final FavoriteService favoriteService;
   private final SearchService searchService;
   private final ImageFileService imageFileService;
+
+  @Override
+  public Post getPost(Long postId, Category category) {
+    return natureRepository.findById(postId)
+        .orElseThrow(() -> new NotFoundException("해당 게시물을 찾을 수 없습니다."));
+  }
+
+  @Override
+  public PostCardDto getPostCardDto(Long postId, Category category, Language language) {
+    PostCardDto postCardDto = natureRepository.findPostCardDto(postId, language);
+    Optional.ofNullable(postCardDto)
+        .orElseThrow(() -> new NotFoundException("해당 게시물을 찾을 수 없습니다."));
+
+    postCardDto.setCategory(PostCategory.NATURE.toString());
+    return postCardDto;
+  }
 
   // 7대 자연 리스트 조회
   public NatureThumbnailDto getNatureList(MemberInfoDto memberInfoDto,
