--- src/main/java/com/jeju/nanaland/domain/restaurant/service/RestaurantCardService.java
@@ -1,29 +1,0 @@
-package com.jeju.nanaland.domain.restaurant.service;
-
-import com.jeju.nanaland.domain.common.data.Language;
-import com.jeju.nanaland.domain.common.data.PostCategory;
-import com.jeju.nanaland.domain.common.dto.PostCardDto;
-import com.jeju.nanaland.domain.common.service.PostCardService;
-import com.jeju.nanaland.domain.restaurant.repository.RestaurantRepository;
-import com.jeju.nanaland.global.exception.NotFoundException;
-import java.util.Optional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class RestaurantCardService implements PostCardService {
-
-  private final RestaurantRepository restaurantRepository;
-
-  @Override
-  public PostCardDto getPostCardDto(Long postId, Language language) {
-
-    PostCardDto postCardDto = restaurantRepository.findPostCardDto(postId, language);
-    Optional.ofNullable(postCardDto)
-        .orElseThrow(() -> new NotFoundException("해당 게시물을 찾을 수 없습니다."));
-
-    postCardDto.setCategory(PostCategory.RESTAURANT.toString());
-    return postCardDto;
-  }
-}
