--- src/test/java/com/jeju/nanaland/domain/experience/service/ExperienceCardServiceTest.java
@@ -1,83 +1,0 @@
-package com.jeju.nanaland.domain.experience.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.nullable;
-import static org.mockito.Mockito.when;
-
-import com.jeju.nanaland.domain.common.data.Category;
-import com.jeju.nanaland.domain.common.data.Language;
-import com.jeju.nanaland.domain.common.dto.ImageFileDto;
-import com.jeju.nanaland.domain.common.dto.PostCardDto;
-import com.jeju.nanaland.domain.common.entity.ImageFile;
-import com.jeju.nanaland.domain.experience.entity.Experience;
-import com.jeju.nanaland.domain.experience.entity.ExperienceTrans;
-import com.jeju.nanaland.domain.experience.repository.ExperienceRepository;
-import java.util.UUID;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.parallel.Execution;
-import org.junit.jupiter.api.parallel.ExecutionMode;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-@Execution(ExecutionMode.CONCURRENT)
-public class ExperienceCardServiceTest {
-
-  @InjectMocks
-  ExperienceCardService experienceCardService;
-
-  @Mock
-  ExperienceRepository experienceRepository;
-
-  @Test
-  @DisplayName("이색체험 카드 정보 조회")
-  void getPostCardDtoTest() {
-    // given
-    ImageFile imageFile = createImageFile();
-    Experience experience = createExperience(imageFile);
-    ExperienceTrans experienceTrans = createExperienceTrans(experience);
-    PostCardDto postCardDto = PostCardDto.builder()
-        .firstImage(new ImageFileDto(imageFile.getOriginUrl(), imageFile.getThumbnailUrl()))
-        .title(experienceTrans.getTitle())
-        .id(experience.getId())
-        .category(Category.EXPERIENCE.toString())
-        .build();
-    when(experienceRepository.findPostCardDto(nullable(Long.class), eq(Language.KOREAN)))
-        .thenReturn(postCardDto);
-
-    // when
-    PostCardDto result = experienceCardService.getPostCardDto(postCardDto.getId(),
-        Language.KOREAN);
-
-    // then
-    assertThat(result.getFirstImage()).isEqualTo(postCardDto.getFirstImage());
-    assertThat(result.getTitle()).isEqualTo(postCardDto.getTitle());
-  }
-
-  ImageFile createImageFile() {
-    return ImageFile.builder()
-        .originUrl(UUID.randomUUID().toString())
-        .thumbnailUrl(UUID.randomUUID().toString())
-        .build();
-  }
-
-  Experience createExperience(ImageFile imageFile) {
-    return Experience.builder()
-        .priority(0L)
-        .firstImageFile(imageFile)
-        .build();
-  }
-
-  ExperienceTrans createExperienceTrans(Experience experience) {
-    return ExperienceTrans.builder()
-        .experience(experience)
-        .language(Language.KOREAN)
-        .title(UUID.randomUUID().toString())
-        .content(UUID.randomUUID().toString())
-        .build();
-  }
-}
