+++ src/test/java/com/jeju/nanaland/domain/festival/service/FestivalServiceTest.java
@@ -1,24 +1,33 @@
 package com.jeju.nanaland.domain.festival.service;
 
 import static com.jeju.nanaland.domain.common.data.Category.FESTIVAL;
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.ArgumentMatchers.nullable;
 import static org.mockito.Mockito.when;
 
+import com.jeju.nanaland.domain.common.data.Category;
 import com.jeju.nanaland.domain.common.data.Language;
+import com.jeju.nanaland.domain.common.dto.ImageFileDto;
+import com.jeju.nanaland.domain.common.dto.PostCardDto;
 import com.jeju.nanaland.domain.common.entity.ImageFile;
+import com.jeju.nanaland.domain.common.entity.Post;
 import com.jeju.nanaland.domain.common.repository.ImageFileRepository;
 import com.jeju.nanaland.domain.common.service.ImageFileService;
 import com.jeju.nanaland.domain.favorite.service.FavoriteService;
 import com.jeju.nanaland.domain.festival.dto.FestivalCompositeDto;
 import com.jeju.nanaland.domain.festival.dto.FestivalResponse.FestivalDetailDto;
+import com.jeju.nanaland.domain.festival.entity.Festival;
+import com.jeju.nanaland.domain.festival.entity.FestivalTrans;
 import com.jeju.nanaland.domain.festival.repository.FestivalRepository;
 import com.jeju.nanaland.domain.member.dto.MemberResponse.MemberInfoDto;
 import com.jeju.nanaland.domain.member.entity.Member;
 import com.jeju.nanaland.domain.member.entity.enums.Provider;
-import com.jeju.nanaland.domain.search.service.SearchService;
 import java.time.LocalDate;
+import java.util.Optional;
+import java.util.UUID;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -40,13 +49,55 @@ class FestivalServiceTest {
   @Mock
   private FavoriteService favoriteService;
   @Mock
-  private SearchService searchService;
-  @Mock
   private ImageFileService imageFileService;
 
 
   @Mock
   private ImageFileRepository imageFileRepository;
+
+  @Test
+  @DisplayName("축제 카드 정보 조회")
+  void getPostCardDtoTest() {
+    // given
+    ImageFile imageFile = createImageFile(1);
+    Festival festival = createFestival(imageFile);
+    FestivalTrans festivalTrans = createFestivalTrans(festival);
+    PostCardDto postCardDto = PostCardDto.builder()
+        .firstImage(new ImageFileDto(imageFile.getOriginUrl(), imageFile.getThumbnailUrl()))
+        .title(festivalTrans.getTitle())
+        .id(festival.getId())
+        .category(Category.FESTIVAL.toString())
+        .build();
+    when(festivalRepository.findPostCardDto(nullable(Long.class), eq(Language.KOREAN)))
+        .thenReturn(postCardDto);
+
+    // when
+    PostCardDto result =
+        festivalService.getPostCardDto(postCardDto.getId(), FESTIVAL, Language.KOREAN);
+
+    // then
+    assertThat(result.getFirstImage()).isEqualTo(postCardDto.getFirstImage());
+    assertThat(result.getTitle()).isEqualTo(postCardDto.getTitle());
+  }
+
+  @Test
+  @DisplayName("축제 Post 조회")
+  void getPostTest() {
+    // given
+    ImageFile imageFile = createImageFile(1);
+    Festival festival = Festival.builder()
+        .priority(0L)
+        .firstImageFile(imageFile)
+        .build();
+    when(festivalRepository.findById(nullable(Long.class)))
+        .thenReturn(Optional.ofNullable(festival));
+
+    // when
+    Post post = festivalService.getPost(1L, Category.FESTIVAL);
+
+    // then
+    assertThat(post.getFirstImageFile()).isEqualTo(imageFile);
+  }
 
   @Test
   @DisplayName("festival 상세 조회에서 기간 국가별 요일")
@@ -121,5 +172,21 @@ class FestivalServiceTest {
         locale, "title", "content", "address", "addressTag", "time", "intro", "fee", startDate,
         endDate,
         "봄");
+  }
+
+  Festival createFestival(ImageFile imageFile) {
+    return Festival.builder()
+        .priority(0L)
+        .firstImageFile(imageFile)
+        .build();
+  }
+
+  FestivalTrans createFestivalTrans(Festival festival) {
+    return FestivalTrans.builder()
+        .festival(festival)
+        .language(Language.KOREAN)
+        .title(UUID.randomUUID().toString())
+        .content(UUID.randomUUID().toString())
+        .build();
   }
 }
