--- src/test/java/com/jeju/nanaland/domain/market/service/MarketCardServiceTest.java
@@ -1,83 +1,0 @@
-package com.jeju.nanaland.domain.market.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.nullable;
-import static org.mockito.Mockito.when;
-
-import com.jeju.nanaland.domain.common.data.Category;
-import com.jeju.nanaland.domain.common.data.Language;
-import com.jeju.nanaland.domain.common.dto.ImageFileDto;
-import com.jeju.nanaland.domain.common.dto.PostCardDto;
-import com.jeju.nanaland.domain.common.entity.ImageFile;
-import com.jeju.nanaland.domain.market.entity.Market;
-import com.jeju.nanaland.domain.market.entity.MarketTrans;
-import com.jeju.nanaland.domain.market.repository.MarketRepository;
-import java.util.UUID;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.parallel.Execution;
-import org.junit.jupiter.api.parallel.ExecutionMode;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-@Execution(ExecutionMode.CONCURRENT)
-public class MarketCardServiceTest {
-
-  @InjectMocks
-  MarketCardService marketCardService;
-
-  @Mock
-  MarketRepository marketRepository;
-
-  @Test
-  @DisplayName("전통시장 카드 정보 조회")
-  void getPostCardDtoTest() {
-    // given
-    ImageFile imageFile = createImageFile();
-    Market market = createMarket(imageFile);
-    MarketTrans marketTrans = createMarketTrans(market);
-    PostCardDto postCardDto = PostCardDto.builder()
-        .firstImage(new ImageFileDto(imageFile.getOriginUrl(), imageFile.getThumbnailUrl()))
-        .title(marketTrans.getTitle())
-        .id(market.getId())
-        .category(Category.MARKET.toString())
-        .build();
-    when(marketRepository.findPostCardDto(nullable(Long.class), eq(Language.KOREAN)))
-        .thenReturn(postCardDto);
-
-    // when
-    PostCardDto result = marketCardService.getPostCardDto(postCardDto.getId(),
-        Language.KOREAN);
-
-    // then
-    assertThat(result.getFirstImage()).isEqualTo(postCardDto.getFirstImage());
-    assertThat(result.getTitle()).isEqualTo(postCardDto.getTitle());
-  }
-
-  ImageFile createImageFile() {
-    return ImageFile.builder()
-        .originUrl(UUID.randomUUID().toString())
-        .thumbnailUrl(UUID.randomUUID().toString())
-        .build();
-  }
-
-  Market createMarket(ImageFile imageFile) {
-    return Market.builder()
-        .priority(0L)
-        .firstImageFile(imageFile)
-        .build();
-  }
-
-  MarketTrans createMarketTrans(Market market) {
-    return MarketTrans.builder()
-        .market(market)
-        .language(Language.KOREAN)
-        .title(UUID.randomUUID().toString())
-        .content(UUID.randomUUID().toString())
-        .build();
-  }
-}
