--- src/test/java/com/jeju/nanaland/domain/nature/service/NatureCardServiceTest.java
@@ -1,83 +1,0 @@
-package com.jeju.nanaland.domain.nature.service;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.ArgumentMatchers.nullable;
-import static org.mockito.Mockito.when;
-
-import com.jeju.nanaland.domain.common.data.Category;
-import com.jeju.nanaland.domain.common.data.Language;
-import com.jeju.nanaland.domain.common.dto.ImageFileDto;
-import com.jeju.nanaland.domain.common.dto.PostCardDto;
-import com.jeju.nanaland.domain.common.entity.ImageFile;
-import com.jeju.nanaland.domain.nature.entity.Nature;
-import com.jeju.nanaland.domain.nature.entity.NatureTrans;
-import com.jeju.nanaland.domain.nature.repository.NatureRepository;
-import java.util.UUID;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.junit.jupiter.api.parallel.Execution;
-import org.junit.jupiter.api.parallel.ExecutionMode;
-import org.mockito.InjectMocks;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
-
-@ExtendWith(MockitoExtension.class)
-@Execution(ExecutionMode.CONCURRENT)
-public class NatureCardServiceTest {
-
-  @InjectMocks
-  NatureCardService natureCardService;
-
-  @Mock
-  NatureRepository natureRepository;
-
-  @Test
-  @DisplayName("전통시장 카드 정보 조회")
-  void getPostCardDtoTest() {
-    // given
-    ImageFile imageFile = createImageFile();
-    Nature nature = createNature(imageFile);
-    NatureTrans natureTrans = createNatureTrans(nature);
-    PostCardDto postCardDto = PostCardDto.builder()
-        .firstImage(new ImageFileDto(imageFile.getOriginUrl(), imageFile.getThumbnailUrl()))
-        .title(natureTrans.getTitle())
-        .id(nature.getId())
-        .category(Category.NATURE.toString())
-        .build();
-    when(natureRepository.findPostCardDto(nullable(Long.class), eq(Language.KOREAN)))
-        .thenReturn(postCardDto);
-
-    // when
-    PostCardDto result = natureCardService.getPostCardDto(postCardDto.getId(),
-        Language.KOREAN);
-
-    // then
-    assertThat(result.getFirstImage()).isEqualTo(postCardDto.getFirstImage());
-    assertThat(result.getTitle()).isEqualTo(postCardDto.getTitle());
-  }
-
-  ImageFile createImageFile() {
-    return ImageFile.builder()
-        .originUrl(UUID.randomUUID().toString())
-        .thumbnailUrl(UUID.randomUUID().toString())
-        .build();
-  }
-
-  Nature createNature(ImageFile imageFile) {
-    return Nature.builder()
-        .priority(0L)
-        .firstImageFile(imageFile)
-        .build();
-  }
-
-  NatureTrans createNatureTrans(Nature nature) {
-    return NatureTrans.builder()
-        .nature(nature)
-        .language(Language.KOREAN)
-        .title(UUID.randomUUID().toString())
-        .content(UUID.randomUUID().toString())
-        .build();
-  }
-}
